Only keep __main__
